Rename classes with overloaded / incorrect names (#359)
This PR renames the following:
* `FSDPRefRayActorBase` -> `FSDPRefRayWorkerBase`: The other training
backends use `Worker`, not `Actor`. Also, `Actor` is overloaded with Ray
actors, so worker is preferred.
* In models.py, `Actor` -> `HFWrappedModel`: Here also `Actor` is
overloaded with Ray actors, and it really just refers to a wrapping of
the underlying HF model. Also, it is not used specifically for the
policy/actor model in RL, but rather used as the base for all models.
* `MegatronPolicy` -> `MegatronWrappedModel`: For Megatron, we use
Megatron models rather than HF models, so this is the Megatron analog to
`HFWrappedModel`. Also, this class is used for non-Policy roles (e.g.,
Reference), so `Policy` is misleading