Add __init__.py to config directory (#418)
This PR adds `__init__.py` file to the `skyrl_train/config` directory to
make `skyrl_train.config` be recognized as a proper package.

This is needed when creating custom configs that use the configs in
`skyrl_train/config` as defaults. This can be done by adding the
`skyrl_train.config` package to Hydra's search path, which only works if
it has an `__init__.py`.

Example custom config:

```yaml
defaults:
  - _self_
  - ppo_base_config

hydra:
  searchpath:
    - pkg://skyrl_train.config

# custom fields below...
```