Fix Megatron batch validation (#393)
When using Megatron, if `cfg.trainer.policy.sequence_parallel_size` is
set to 1, `policy_dp_size` will always be the world size. However, if
any of PP, TP, CP is positive, this does not hold. This PR added the
correct way to calculate `policy_dp_size` and `ref_dp_size` when using
Megatron.

---------

Co-authored-by: Eric Tang <erictang000@gmail.com>