[tx] Add MultiLoRA MoE layer (#432)
This PR implements MultiLoRA for the MoE layer. It is implemented by
routing tokens to the (num_experts, num_adapters) dimensions of the
lora_A and lora_B matrices, much in the same way that tokens are routed
to the num_experts dimension of the experts matrices for normal MoE
layers. We also add a test against merging each LoRA adapter
individually into the base model to ensure numerical correctness of the
implementation.