[fix] Resolve timeout and cleanup issues in GPU CI pipeline (#483)
The primary change was to remove the bespoke ways that different tests
are setting up and tearing down resources (e.g,. `ray.init()`,
`ray.shutdown()`), and in this PR we instead prefer the use of
`ray_init_fixture`.

This also fixes an indent error in the creation of dp>1 Ray-wrapped
inference engines.