[tx] Adds a  CheckpointsDB model to track checkpoints and their status (#486)
This PR implements a way to keep track of checkpoints in the database.
Additionally, the implementation ensures that checkpoint IDs are unique
and an error is throws when a checkpoint ID is duplicated. Let me know
if we need to handle this edge case differently. Next steps after this
gets approval will be to add endpoints to support for listing and
filtering checkpoints as well as access controls once an auth support is
in place.

---------

Co-authored-by: Philipp Moritz <pcmoritz@gmail.com>