[megatron][bug fix] disable megatron spawn without persistent dist-checkpointing backend (#565)
# Overview
closes #556
## Bug
In the constructor for MegatronStrategy, we have the following code:

 ```python
# NOTE: Set Megatron dist checkpoint async backend to persistent to
avoid `os.fork()`-ing
# short-lived background workers, which does not work well with Ray.
ckpt_base.async_calls = AsyncCallsQueue(persistent=True)
```

This was added in #357 to change the backend to be persistent, in order to avoid forking during megatron dist checkpointing. However, this caused it to not be possible to offload optimizer or model memory to cpu following a checkpointing step (maybe dangling tensor references in the dist checkpointing backend preventing memory from being freed).

## Solution
We can get around this by closing the `AsyncCallsQueue` object, and resetting it on each call to `save_checkpoints`.

I've confirmed that checkpointing still runs (both single and multi-node) without issues, and updated the checkpointing gpu ci test to make sure that cpu memory can still be offloaded after checkpointing.