[tx] Make token weights optional (#584)
If the token weights are not provided by the user, they will be filled
in to all be 1.0. Also fixes the predicted tokens and token weights in
the test (doesn't impact the test, but it is good to fix it).

This is needed for https://github.com/NovaSky-AI/SkyRL/issues/535