Fix: Megatron Autograd Warning for Broadcast Kernel (#588)
#370 The broadcast kernel from torch.distributed is being called
implicitly during `ppo_train` backprop. Training with Megatron leads to
the following autograd warning:

`[36m(MegatronPolicyWorkerBase pid=26429, ip=10.76.2.247)[0m
/workspace/SkyRL/skyrl-train/.venv/lib/python3.12/site-packages/torch/autograd/graph.py:824:
UserWarning: c10d::broadcast_: an autograd kernel was not registered to
the Autograd key(s) but we are trying to backprop through it. This may
lead to silently incorrect behavior. This behavior is deprecated and
will be removed in a future version of PyTorch. If your operator is
differentiable, please ensure you have registered an autograd kernel to
the correct Autograd key (e.g. DispatchKey::Autograd,
DispatchKey::CompositeImplicitAutograd). If your operator is not
differentiable, or to squash this warning and use the previous behavior,
please register torch::CppFunction::makeFallthrough() to
DispatchKey::Autograd. (Triggered internally at
/pytorch/torch/csrc/autograd/autograd_not_implemented_fallback.cpp:62.)
`

To prevent the autograd engine from registering broadcast, we should
only be using this kernel with `torch.no_grad()`. We explicitly set
broadcast to be called with this decorator wrapped around it to override
the function behaviour.

---------