[skyrl-train][Examples] Support truncated importance sampling for `StepWiseGenerator` (#570)
# What does this PR do?

Adds support for truncated importance sampling (TIS ) with the step wise
example. This means that TIS can now be used with multi-turn training

The implementation is mostly straightforward. One thing to note is that
we append an EOS token in the agent loop when stop tokens are passed.
This is a new token that was not generated by the model and thus
logprobs are not available - thus this token is masked out.

---------

Signed-off-by: SumanthRH <sumanthrh@anyscale.com>
Signed-off-by: SumanthRH <sumanthrh99@gmail.com>