[train][utils] Aggregate rollout metrics and validate output in concat GeneratorOutput (#620)
This PR makes two changes to the util `concatenate_generator_outputs()`
which is mainly used in evaluation. But it will soon be used in other
cases (e.g. fully async training), so the following two functionalities
would be helpful:
- Aggregate rollout metrics (those that can be deduced from response_ids
and rewards)
- Validate generator output

Note that we also remove `GeneratorInput` to the validation util, and
directly require a `num_prompts` instead, since in some cases the
`GeneratorInput` is not accessible.

Changed unit tests accordingly