[skyrl-train][Fix] Fix FSDP1 module wrap policy for `HFModelWrapper` (#654)
# What does this PR do?

The wrap policy was incorrect for FSDP1 - we were directly passing the
wrapped model instead of using `model.model`. This lead to only the full
model being wrapped as one FSDP unit which is pretty terrible.

This PR fixes the issue and adds a test. Test fails before this fix

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>