[skyrl-train] Upgrade Ray to 2.51.1 (#633)
# What does this PR do?

Upgrades Ray to 2.51. Closes #632 


Note that this is a hard upgrade. This is because SkyRL is changing so
fast that the docker image only packages the dependencies and not a
versioned package of the repo. Further, the typical workflow is that
users add to the SkyRL repo to add their own example, etc.

SkyRL can still be run with ray < 2.5.1, and that section of the docs is
retained as is.

We will improve this dependency story in the future.

TODO:
- [x] Publish new docker image
- [x] Update docs

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>