[skyrl-train][megatron] Remove use of PYTHONPATH for getting around transformer-engine installation (#697)
# Overview
Previously, we were setting the `PYTHONPATH` manually in order to point
uv to the cached installation of transformer-engine in the docker image.
This was hacky and could cause issues with dependencies. Now with the
upgraded uv version allowing us to install transformer-engine with
matched runtime with torch, and adding ninja to the base dependencies
for faster build, this hack is no longer needed. Leaving the code path
for backwards compatibility, but removing it from scripts and docs.