[tx] Do not log path of Python file and line number (#782)
Currently, that path and line number of the python statement generating
the log is logged, in a right aligned fashion. While this sounds useful,
it practice it isn't really and makes working with the logs somewhat
annoying (since there is a lot of white space and dependence on the
terminal size).