[train][step-wise] 2/N - Refactor agent loop and support step-wise training with default SkyRLGymGenerator (#774)
# What does this PR do?


WIP PR to refactor agent loop and support step wise training in the
default `SkyRLGymGenerator`. With this PR, we no longer need a separate
`StepWiseGenerator`.

This PR also refactors the agent loop and introduces new dataclasses
`TurnOutput` and `AgentLoopState` to better manage state in the agent
loop. There are breaking changes to the interface for
`SkyRLGymGenerator.agent_loop`

Note that the existing `retokenize_chat_history` +
`custom_chat_template` codepath is kept intact. We should consider
deprecating this codepath since that was primarly for computing correct
loss masks for thinking models like Qwen. We can use step wise training
as an alternative.

TODO: 
- [x] Cleanup

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>