[AsyncRL][N/N] Support TIS to fully async and multi-turn SkyRLGymGenerator (#799)
This PR supports TIS for fully async RL and thus conclude the support
for fully async RL in SkyRL.

Besides, with this PR, we can use fully async for any task supported in
`SkyRLGymGenerator`, and hence we can remove the
`SkyRLGymHTTPGenerator`.

We also support TIS for mutli-turn generation in `SkyRLGymGenerator`
(async or not) -- this was already supported, but just needed some
config tuning.

Separately, we fix a bug in computing `expected_consumed_in_epoch` when
we are at the boundary of an epoch.

### Tests

We validate multi-turn TIS with Search r1, with the curve comparison
below. We indeed see some small improvement with TIS.
<img width="1676" height="285" alt="image"
src="https://github.com/user-attachments/assets/fc0ea1cc-1775-4b9f-908d-0f6088ba6c17"
/>

Also tested fully async DAPO w/ TIS, will add the script and results in
a follow up PR.